[status] add checks for pending and viewable
`isPending` checks whether we are waiting for data to enter the store.
This will return `false` as long as there is data to be returned,
regardless of it's state.

`isViewable` checks whether there is data in the store that can be
displayed.
This will return `true` as long as there is data to be returned,
regardless of it's state.