fix(vuex): prevent updeep from freezing state
When using update on state with an empty update object the state passed
in might be freezed.
This will make any attempt to `assign()` properties fail due to the
initial state being frozen. Effectively making the vuex state fully
immutable.

This is solved by passing in a very superficial clone into any operation
passed to the `apply()` function.